re-organized in verbose modules
