+++ doc/conf.py
@@ -117,15 +117,15 @@ todo_include_todos = True
 html_theme = "sphinx_rtd_theme"
 
 html_theme_options = {
-    'collapse_navigation': False,
-    'display_version': True,
-    'logo_only': True,
-    'navigation_depth': 2,
+    "collapse_navigation": False,
+    "display_version": True,
+    "logo_only": True,
+    "navigation_depth": 2,
 }
 
-html_logo = '_static/img/omc_logo.svg'
-html_static_path = ['_static']
-html_context = {'css_files': ['_static/css/custom.css']}
+html_logo = "_static/img/omc_logo.svg"
+html_static_path = ["_static"]
+html_context = {"css_files": ["_static/css/custom.css"]}
 
 smartquotes_action = "qe"  # renders only quotes and ellipses (...) but not dashes (option: D)
