+++ tfs/errors.py
@@ -1,0 +1,12 @@
+"""
+Errors
+-------------------
+
+Errors that can be raised during the handling of **TFS** files.
+"""
+
+
+class TfsFormatError(Exception):
+    """Raised when a wrong format is detected in the TFS file."""
+
+    pass
